test: fold wallet subscriptions into unshielded
`wallet-subscriptions.test.ts` held no shielded and no dust coverage:
every live assertion went through the `unshieldedTransactions(address:)`
subscription, so it was unshielded-transaction coverage sitting under a
consumer-shaped filename. Move its three live describes, the
`validateCrossWalletTransaction` helper and the four backlog `test.todo`
entries into `unshielded-transactions.test.ts` and delete the file. The
merged layout mirrors the one `shielded-transactions.test.ts` already
uses for its subscription block.

The second destination wallet is now subscribed by the file-level
fixture, and the moved multi-destination block runs after the existing
one without clearing the shared event buffers, which the existing tests
compare against baselines captured in `beforeAll`.

Two defects in the moved assertions are fixed rather than carried over.
The empty-wallet test called `every()` on a possibly empty array, which
is vacuously true, so it passed even when the indexer emitted nothing;
it now requires at least one event. The value-conservation and
`spentAtTransaction` checks were wrapped in an `if (spent)` guard that
made them disappear whenever `spentUtxos` was empty; the spent UTXO is
now asserted to be present.

The two subscription tests for the 1 STAR transfer matched events on
`owner` alone. Every e2e file spends from the same funding wallet into
the same destination seed and the files run in parallel workers, so a
transfer submitted by another file could satisfy them. Both now match
on the submitted transaction hash.

----
